limited access with different ranks

+++ private/controllers/Classes.php
@@ -17,7 +17,7 @@ class Classes extends Controller
         $data = $classes->query("select * from classes where school_id = :school_id order by id desc", ['school_id' => $school_id]);
 
         $crumbs[] = ['Tableau de bord', ''];
-        $crumbs[] = ['Écoles', 'classes'];
+        $crumbs[] = ['Cours', 'classes'];
 
         $this->view('classes', [
             'rows' => $data,
@@ -47,7 +47,7 @@ class Classes extends Controller
             }
         }
         $crumbs[] = ['Tableau de bord', ''];
-        $crumbs[] = ['École', 'classes'];
+        $crumbs[] = ['Cours', 'classes'];
         $crumbs[] = ['Ajouter', 'classes/add'];
 
         $this->view(
@@ -64,9 +64,10 @@ class Classes extends Controller
             $this->redirect('login');
         }
         $classes = new CLasses_model();
-        $errors = array();
 
-        if (count($_POST) > 0) {
+        $errors = array();
+        if (count($_POST) > 0 && Auth::access('Enseignant.e') && Auth::i_own_content($row)) {
+
             if ($classes->validate($_POST)) {
 
                 $classes->update($id, $_POST);
@@ -79,45 +80,54 @@ class Classes extends Controller
         $row = $classes->where('id', $id);
 
         $crumbs[] = ['Tableau de bord', ''];
-        $crumbs[] = ['École', 'classes'];
+        $crumbs[] = ['Cours', 'classes'];
         $crumbs[] = ['Modifier', 'classes/edit'];
 
-        $this->view(
-            'classes.edit',
-            [
-                'row' => $row,
-                'errors' => $errors,
-                'crumbs' => $crumbs,
+        if (Auth::access('Enseignant.e') && Auth::i_own_content($row)) {
+            $this->view(
+                'classes.edit',
+                [
+                    'row' => $row,
+                    'errors' => $errors,
+                    'crumbs' => $crumbs,
 
-            ]
-        );
+                ]
+            );
+        } else {
+            $this->view('access-denied');
+        }
     }
     public function delete($id = null)
     {
         if (!Auth::logged_in()) {
             $this->redirect('login');
         }
+
+
         $classes = new Classes_model();
         $errors = array();
 
-        if (count($_POST) > 0) {
+        if (count($_POST) > 0 && Auth::access('Enseignant.e') && Auth::i_own_content($row)) {
             $classes->delete($id);
             $this->redirect('classes');
         }
-
         $row = $classes->where('id', $id);
 
         $crumbs[] = ['Tableau de bord', ''];
-        $crumbs[] = ['Écoles', 'classes'];
+        $crumbs[] = ['Cours', 'classes'];
         $crumbs[] = ['Supprimer', 'classes/delete'];
 
-        $this->view(
-            'classes.delete',
-            [
-                'row' => $row,
-                'crumbs' => $crumbs,
+        if (Auth::access('Enseignant.e') && Auth::i_own_content($row)) {
+            $this->view(
+                'classes.delete',
+                [
+                    'row' => $row,
+                    'crumbs' => $crumbs,
 
-            ]
-        );
+                ]
+            );
+        } else {
+            $this->view('access-denied');
+        }
     }
 }
