+++ private/controllers/Schools.php
@@ -18,11 +18,15 @@ class Schools extends Controller
         $crumbs[] = ['Tableau de bord', ''];
         $crumbs[] = ['Écoles', 'schools'];
 
-        $this->view('schools', [
-            'rows' => $data,
-            'crumbs' => $crumbs,
+        if (Auth::access('Super Administrateur.trice')) {
+            $this->view('schools', [
+                'rows' => $data,
+                'crumbs' => $crumbs,
 
-        ]);
+            ]);
+        } else {
+            $this->view('access-denied');
+        }
     }
     public function add()
     {
@@ -31,7 +35,7 @@ class Schools extends Controller
         }
         $errors = array();
 
-        if (count($_POST) > 0) {
+        if (count($_POST) > 0 && Auth::access('Super Administrateur.trice')) {
             $school = new School();
             if ($school->validate($_POST)) {
 
@@ -49,13 +53,17 @@ class Schools extends Controller
         $crumbs[] = ['Écoles', 'schools'];
         $crumbs[] = ['Ajouter', 'schools/add'];
 
-        $this->view(
-            'schools.add',
-            [
-                'errors' => $errors,
-                'crumbs' => $crumbs,
-            ]
-        );
+        if (Auth::access('Super Administrateur.trice')) {
+            $this->view(
+                'schools.add',
+                [
+                    'errors' => $errors,
+                    'crumbs' => $crumbs,
+                ]
+            );
+        } else {
+            $this->view('access-denied');
+        }
     }
     public function edit($id = null)
     {
@@ -65,7 +73,7 @@ class Schools extends Controller
         $school = new School();
         $errors = array();
 
-        if (count($_POST) > 0) {
+        if (count($_POST) > 0 && Auth::access('Super Administrateur.trice')) {
             if ($school->validate($_POST)) {
 
                 $school->update($id, $_POST);
@@ -81,15 +89,18 @@ class Schools extends Controller
         $crumbs[] = ['Écoles', 'schools'];
         $crumbs[] = ['Modifier', 'schools/edit'];
 
-        $this->view(
-            'schools.edit',
-            [
-                'row' => $row,
-                'errors' => $errors,
-                'crumbs' => $crumbs,
-
-            ]
-        );
+        if (Auth::access('Super Administrateur.trice')) {
+            $this->view(
+                'schools.edit',
+                [
+                    'row' => $row,
+                    'errors' => $errors,
+                    'crumbs' => $crumbs,
+                ]
+            );
+        } else {
+            $this->view('access-denied');
+        }
     }
     public function delete($id = null)
     {
@@ -98,7 +109,7 @@ class Schools extends Controller
         }
         $school = new School();
         $errors = array();
-        if (count($_POST) > 0) {
+        if (count($_POST) > 0 && Auth::access('Super Administrateur.trice')) {
             $school->delete($id);
             $this->redirect('schools');
         }
@@ -108,13 +119,16 @@ class Schools extends Controller
         $crumbs[] = ['Écoles', 'schools'];
         $crumbs[] = ['Supprimer', 'schools/delete'];
 
-        $this->view(
-            'schools.delete',
-            [
-                'row' => $row,
-                'crumbs' => $crumbs,
-
-            ]
-        );
+        if (Auth::access('Super Administrateur.trice')) {
+            $this->view(
+                'schools.delete',
+                [
+                    'row' => $row,
+                    'crumbs' => $crumbs,
+                ]
+            );
+        } else {
+            $this->view('access-denied');
+        }
     }
 }
